MotionDesign : PropertyAnimator - Cleaning public facing API
#rb Juan.Portillo

#changelist validated
#virtualized

[CL 31520073 by loris rossi in 5.4 branch]

+++ Engine/Plugins/Experimental/PropertyAnimatorCore/Source/PropertyAnimatorCore/Private/TimeSources/PropertyAnimatorCoreManualTimeSource.cpp
@@ -2,7 +2,7 @@
 
 #include "TimeSources/PropertyAnimatorCoreManualTimeSource.h"
 
-void UPropertyAnimatorCoreManualTimeSource::SetCustomTime(float InTime)
+void UPropertyAnimatorCoreManualTimeSource::SetCustomTime(double InTime)
 {
 	if (FMath::IsNearlyEqual(InTime, CustomTime))
 	{
